refactor: move blog and changelog loaders into enterprise
