+++ app/lib/glossia/changelog.ex
@@ -7,3 +7,12 @@ defmodule Glossia.Changelog do
 
   def all_entries, do: Glossia.Extensions.changelog().all_entries()
 end
+
+defmodule Glossia.Changelog.Empty do
+  @moduledoc false
+
+  @behaviour Glossia.Changelog.Source
+
+  @impl true
+  def all_entries, do: []
+end
